+++ config/typedoc.json
@@ -8,9 +8,9 @@
         "../src/services/settings_manager.ts",
         "../src/services/style_service.ts",
         "../src/services/notification_service.ts",
-        "../src/domain/models/color.ts",
-        "../src/domain/models/settings_keys.ts",
-        "../src/domain/types/ui.ts",
+        "../src/models/color.ts",
+        "../src/models/settings_keys.ts",
+        "../src/types/ui.ts",
         "../src/utils/color/color_utils.ts",
         "../src/utils/color/accent_color_utils.ts",
         "../src/utils/parse_utils.ts",
