+++ src/tests/unit/utils/style_utils.test.ts
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-import { Color } from "../../../domain/models/color.js";
+import { Color } from "../../../models/color.js";
 import { buildStyles } from "../../../utils/style/style_utils.js";
 
 describe("buildStyles", () => {
